AsyncProcessAvContent: Adjust logic when fetching new content
In some cases, the M3U playlist might be empty when fetching new
content. (Caused by something on the external API)  This either
deletes the cache's contents or leave a section loading indefinitely.

If the cache does exist, don't touch anything in it and show an error
message if it doesn't.

+++ library/src/main/java/com/zaclimon/xipl/ui/vod/VodTvSectionFragment.java
@@ -261,15 +261,18 @@ public abstract class VodTvSectionFragment extends RowsSupportFragment {
 
                 if (!isCancelled()) {
                     final List<AvContent> avContents = AvContentUtil.getAvContentsList(catchupInputStream, VodTvSectionFragment.this.getClass().getSimpleName());
-
                     if (avContents.size() != persistedSize && persistedSize == 0) {
+                        // Case where the list is being loaded for the first time.
                         getContentPersistence().insert(avContents);
-                    } else if (avContents.size() != persistedSize) {
+                    } else if (avContents.size() != persistedSize && persistedSize != 0) {
+                        // Case where the contents list have been modified upstream.
                         getContentPersistence().deleteCategory(VodTvSectionFragment.this.getClass().getSimpleName());
                         getContentPersistence().insert(avContents);
+                    } else {
+                        // Case where the content list might be empty. In that case, don't touch the cache.
+                        return (false);
                     }
                 }
-
                 return (true);
             } catch (IOException io) {
                 return (false);
